removed all legacy CEHR-BERT code using tensorflow and keras
