+++ constraints.txt
@@ -1,0 +1,1 @@
+xformers==0 # blocks installation
