+++ pyproject.toml
@@ -46,12 +46,8 @@ dependencies = [
     "PyYAML==6.0.1",
     "scikit-learn==1.4.0",
     "scipy==1.12.0",
-    "tensorflow>=2.15.0",
-    "keras==2.15.0",
-    "tensorflow-metal>=1.1.0; sys_platform == 'darwin'", # macOS only
-    "tensorflow-datasets>=4.5.2",
     "tqdm>=4.66.1",
-    "torch==2.4.0",
+    "torch>=2.4.0",
     "tokenizers>=0.19.0",
     "transformers>=4.41.0, <= 4.45.0",
     "accelerate>=0.31.0",
