--- src/cehrbert/config/output_names.py
@@ -1,9 +1,0 @@
-parquet_data_path = "patient_sequence"
-qualified_concept_list_path = "qualified_concept_list"
-time_attention_model_path = "time_aware_model.h5"
-bert_model_validation_path = "bert_model.h5"
-mortality_data_path = "mortality"
-heart_failure_data_path = "heart_failure"
-hospitalization_data_path = "hospitalization"
-information_content_data_path = "information_content"
-concept_similarity_path = "concept_similarity"
