--- src/cehrbert/models/loss_schedulers.py
@@ -1,56 +1,0 @@
-import math
-
-
-class CosineLRSchedule:
-    """Cosine annealing with warm restarts, described in paper.
-
-    "SGDR: stochastic gradient descent with warm restarts"
-    https://arxiv.org/abs/1608.03983
-
-    Changes the learning rate, oscillating it between `lr_high` and `lr_low`.
-    It takes `period` epochs for the learning rate to drop to its very minimum,
-    after which it quickly returns back to `lr_high` (resets) and everything
-    starts over again.
-
-    With every reset:
-        * the period grows, multiplied by factor `period_mult`
-        * the maximum learning rate drops proportionally to `high_lr_mult`
-
-    This class is supposed to be used with
-    `keras.callbacks.LearningRateScheduler`.
-    """
-
-    def __init__(
-        self,
-        lr_high: float,
-        lr_low: float,
-        initial_period: int = 50,
-        period_mult: float = 2,
-        high_lr_mult: float = 0.97,
-    ):
-        self._lr_high = lr_high
-        self._lr_low = lr_low
-        self._initial_period = initial_period
-        self._period_mult = period_mult
-        self._high_lr_mult = high_lr_mult
-
-    def __call__(self, epoch, lr):
-        return self.get_lr_for_epoch(epoch)
-
-    def get_lr_for_epoch(self, epoch):
-        assert epoch >= 0
-        t_cur = 0
-        lr_max = self._lr_high
-        period = self._initial_period
-        result = lr_max
-        for i in range(epoch + 1):
-            if i == epoch:  # last iteration
-                result = self._lr_low + 0.5 * (lr_max - self._lr_low) * (1 + math.cos(math.pi * t_cur / period))
-            else:
-                if t_cur == period:
-                    period *= self._period_mult
-                    lr_max *= self._high_lr_mult
-                    t_cur = 0
-                else:
-                    t_cur += 1
-        return result
