--- src/cehrbert/utils/logging_utils.py
@@ -1,16 +1,0 @@
-import logging
-import sys
-
-
-def add_console_logging():
-    root = logging.getLogger()
-    root.setLevel(logging.INFO)
-
-    handler = logging.StreamHandler(sys.stdout)
-    handler.setLevel(logging.DEBUG)
-    formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
-    handler.setFormatter(formatter)
-    root.addHandler(handler)
-
-
-add_console_logging()
